updated Audio vector to uint16

+++ include/audio/audio.hh
@@ -1,8 +1,9 @@
 #ifndef AUDIO_AUDIO_HH
 #define AUDIO_AUDIO_HH
 
+#include <cstdint>
 #include <vector>
 
-using Audio = std::vector<float>;
+using Audio = std::vector<uint16_t>;
 
 #endif // AUDIO_AUDIO_HH
